Made changes to minimize the errors in files of login folder and created route in main.dart file
