+++ lib/main.dart
@@ -1,4 +1,5 @@
-import 'package:ap/login/loginscreen.dart';
+import 'package:ap/Register/register_page.dart';
+import 'package:ap/login/login_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -14,7 +15,10 @@ class MyApp extends StatelessWidget {
       title: 'Login page',
       theme: ThemeData.light(),
       debugShowCheckedModeBanner: false,
-      home: const LoginScreen(),
+      routes: {
+        '/': (context) => const LoginScreen(),
+        '/register': (context) => const RegisterScreen()
+      },
     );
   }
 }
